LayoutTests/imported/w3c/web-platform-tests/intersection-observer/transformed-iframe-003-same-origin-expected.txt



PASS Target is not visible as it's rotated perpendicularly to viewport
PASS Scroll down a bit, target is still not visible
